Move google config
